refactor: introduce application layer in Interop
